Still gyroscope problems
I suspect we might have to ask for location permission on the phone itself (for some reason) Will research a bit further. Also, we need to find out if we want the bal to actually roll (less control) or if we want it to just move with the tilt/joystick, like a "normal" third person game. The latter I think works best for most gameplay.