Levels are now ready to be prototyped.
Idea is to make at most 3 levels, showing how the game would increase in difficulty as the game progress. Level "1" and Level "50" are the two highest priorities in this regard, and the level between- may be good to have just to show the speed in which the levels are progressively getting more difficult, but also -may not be as important as "1" and "50" for now.